windows warnings

+++ src/Brain/BrainDribble.hpp
@@ -1,9 +1,9 @@
 #pragma once
+#include <iostream>
+
 #include "BrainState.hpp"
 #include "Engine/Compass.hpp"
 #include "Engine/Vector.hpp"
-
-#include <iostream>
 
 class Brain;
 //
@@ -21,8 +21,7 @@ class DribblePattern {
   virtual bool changeDirection() = 0;
 
  protected:
-  int count=0;
-
+  int count = 0;
 };
 //
 //
@@ -34,7 +33,7 @@ class DribblePatternRandom : public DribblePattern {
   //
   Engine::Compass nextDirection(Engine::Compass in_current_dir) override {
     Engine::Vector3 v = in_current_dir.toVector();
-    v.rotate(rand() % 2 == 1 ? 45 : -45);
+    v.rotate(static_cast<float>(rand() % 2 == 1 ? 45 : -45));
     Engine::Compass dir(v);
     return dir;
   }
@@ -43,8 +42,8 @@ class DribblePatternRandom : public DribblePattern {
   //
   bool changeDirection() override {
     static const int min = 60;
-    if(++count > min){
-      if(rand() % 100 > 45){
+    if (++count > min) {
+      if (rand() % 100 > 45) {
         count = 0;
         return true;
       }
